login logout feature added

+++ StudentInformationSystem/StudentInformationSystem/EditStudentform.cs
@@ -15,20 +15,14 @@ namespace StudentInformationSystem
     public partial class EditStudentform : Form
     {
         private StudentDetailsModel model;
+        private string registerno;
+        private readonly Form PreviousWindow;
 
-        public EditStudentform (string regno)
+        public EditStudentform (string regno, Form previousWindow)
         {
             InitializeComponent();
-            EditStudentServices edit=new EditStudentServices();
-            model= edit.StudentGetByRegisterNumber(regno);
-            nametxtbox.Text = model.StudentName;
-            RegNoTxtbox.Text = model.RegisterNumber;
-            emailidtxtbox.Text = model.Emailid;
-            phnnotxtbox.Text = model.PhoneNumber.ToString();
-            departmenttxtbox.Text = model.Department;
-            branchtxtbox.Text = model.Course;
-            staffadvisortxtbox.Text = model.StaffAdvisor;
-            currentsemtxtbox.Text = model.CurrentSemester.ToString();
+            registerno = regno;
+            this.PreviousWindow = previousWindow;
 
         }
         private void updatebtn_Click(object sender, EventArgs e)
@@ -38,8 +32,18 @@ namespace StudentInformationSystem
             studentmodel.RegisterNumber= RegNoTxtbox.Text;
             studentmodel.Emailid= emailidtxtbox.Text;
             studentmodel.PhoneNumber = Convert.ToInt64(phnnotxtbox.Text);
-            studentmodel.Department= departmenttxtbox.Text;
-            studentmodel.Course= branchtxtbox.Text;
+            studentmodel.Department= depcomboBox1.SelectedItem.ToString();
+            if (depcomboBox1.SelectedItem.ToString() == "Computer Science")
+            {
+                coursecomboBox2.Items.Add("Computer Science and Engineering");
+                coursecomboBox2.Items.Add("BSC computer Science");
+            }
+            if (depcomboBox1.SelectedItem.ToString() == "Electronics")
+            {
+                coursecomboBox2.Items.Add("Electronics and communication Engineering");
+                coursecomboBox2.Items.Add("Electrical and Electronics Engineering");
+            }
+            studentmodel.Course= coursecomboBox2.SelectedItem.ToString();
             studentmodel.StaffAdvisor= staffadvisortxtbox.Text;
             studentmodel.CurrentSemester=Convert.ToInt32(currentsemtxtbox.Text);
             EditStudentServices studentServices = new EditStudentServices();
@@ -50,16 +54,58 @@ namespace StudentInformationSystem
 
         private void AddMarksBtn_Click(object sender, EventArgs e)
         {
-            new addmark(RegNoTxtbox.Text,currentsemtxtbox.Text).Show();
+            new addmark(RegNoTxtbox.Text,currentsemtxtbox.Text,this).Show();
         }
 
         private void Deletestudentbtn_Click(object sender, EventArgs e)
         {
-            DeleteStudentService studentService = new DeleteStudentService();
-            int i= studentService.DeleteStudentStudentMarksTable(RegNoTxtbox.Text,emailidtxtbox.Text);
-            if (i != 0)
-                MessageBox.Show($"Student with Register Number {RegNoTxtbox.Text} Deleted!!!!!!!");
+            DialogResult res = MessageBox.Show("Are you sure you want to Delete this Student", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
+            if (res==DialogResult.Yes)
+            {
+                DeleteStudentService studentService = new DeleteStudentService();
+                int i = studentService.DeleteStudentStudentMarksTable(RegNoTxtbox.Text, emailidtxtbox.Text);
+
+                if (i != 0)
+                    MessageBox.Show($"Student with Register Number {RegNoTxtbox.Text} Deleted!");
+            }
             this.Close();
+        }
+
+        private void EditStudentform_Load(object sender, EventArgs e)
+        {
+            EditStudentServices edit = new EditStudentServices();
+            model = edit.StudentGetByRegisterNumber(registerno);
+            if(model == null)
+            {
+                MessageBox.Show("Incorrect Sudent Register number");
+                Close();
+                return;
+            }
+            nametxtbox.Text = model.StudentName;
+            RegNoTxtbox.Text = model.RegisterNumber;
+            emailidtxtbox.Text = model.Emailid;
+            phnnotxtbox.Text = model.PhoneNumber.ToString();
+            depcomboBox1.SelectedItem = model.Department;
+            if (depcomboBox1.SelectedItem.ToString() == "Computer Science")
+            {
+                coursecomboBox2.Items.Add("Computer Science and Engineering");
+                coursecomboBox2.Items.Add("BSC computer Science");
+            }
+            if (depcomboBox1.SelectedItem.ToString() == "Electronics")
+            {
+                coursecomboBox2.Items.Add("Electronics and communication Engineering");
+                coursecomboBox2.Items.Add("Electrical and Electronics Engineering");
+            }
+            coursecomboBox2.SelectedItem = model.Course;
+            staffadvisortxtbox.Text = model.StaffAdvisor;
+            currentsemtxtbox.Text = model.CurrentSemester.ToString();
+
+        }
+
+        private void HOME_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
+        {
+            PreviousWindow.Show();
+            Hide();
         }
     }
 }
